smart-query-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.1.0
pymysql==1.1.0
python-dotenv==1.0.0
